Use int instead of float for texture slot index

+++ Elm/content/shaders/generic_2d.glsl
@@ -4,13 +4,13 @@
 layout(location = 0) in vec3 a_position;
 layout(location = 1) in vec2 a_uv;
 layout(location = 2) in vec4 a_color;
-layout(location = 3) in float a_texture_slot;
+layout(location = 3) in int a_texture_slot;
 
 uniform mat4 u_view_projection;
 
 out vec2 v_uv;
 out vec4 v_color;
-out flat float v_texture_slot;
+out flat int v_texture_slot;
 
 void main()
 {
@@ -27,11 +27,11 @@ layout(location = 0) out vec4 o_color;
 
 in vec2 v_uv;
 in vec4 v_color;
-in flat float v_texture_slot;
+in flat int v_texture_slot;
 
 uniform sampler2D u_textures[32];
 
 void main()
 {
-	o_color = texture(u_textures[int(v_texture_slot)], v_uv) * v_color;
+	o_color = texture(u_textures[v_texture_slot], v_uv) * v_color;
 }
